Remove directory-sep-char, since the deprecated variable was removed from Emacs 24.2
I also removed the global variables clearcase-pname-sep-regexp and
clearcase-non-pname-sep-regexp, no reason to pollute the namespace with
simple constants.

I found this post from Eli Zaretskii to justify the change
<http://lists.gnu.org/archive/html/emacs-devel/2001-11/msg00984.html>:
The problem is that the assumption that `/' is a directory separator
is built into the Emacs code in too many places.  We solved this on
the C level (see the IS_DIRECTORY_SEP macro), but many Lisp files have
`/' hard-wired into them as well.  What's more, quite a few of those
places have regexps set up on the assumption that `/' is the
separator, and other places use the backslashes in file names for
conflicting purposes (such as quoting).

Making all this support directory-sep-char is a lifetime's work that
no one will probably ever invest.

The MS-DOS port never supported directory-sep-char (it always behaves
as if its value were `/').  The Windows port heroically tried to
support it for a few years, but Andrew and Jason came to the
conclusion that it couldn't be supported in a consistent manner (even
if you set it to `\', the forward slash keeps creeping in in all kinds
of places, and some functions break).  So we decided to be honest and
drop it.

Note that removal of this variable doesn't mean we don't support
native DOS/Windows file name.  We do support them, but Emacs simply
converts the file names to the forward-slashes form at the earliest
possible opportunity.